+++ tests/test_vtes.py
@@ -165,9 +165,10 @@ May 8th 2010
 10 players
 Rudolf Scholz
 
+-- 4VP in final
+
 Deck Name: The Storage Procurers
 Description:
-4vp in final
 Allies with Flash Grenades to keep troubles at bay.
 Storage Annex for card efficiency and a structured hand. Weenies and
 Midcaps with Obfuscate and/or Dominate to oust via Conditionings and